Create UI to be used by all architectures [AAG-23] (#21)
If you want to create crossed out TextView - you should do this through span(kotlin). Example:
textView.paintFlags = textView.paintFlags or STRIKE_THRU_TEXT_FLAG

However, Toast and Snackbar do not apply to the UI. For everything to work, just apply the main theme from the UI module. Also, there was no dialogue design, so I've done a simple form.
The AlertDialog wasn't added to the UI module because I am not sure if this is the area of responsibility of the UI module.